# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Merced's Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface AuthInput {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  hasAccount: boolean;
}

export interface TradingAccount {
  id: number;
  platformName: string;
  accountId: string;
  capital: number;
  profitTarget: number;
  riskLevel: string;
  timeframe: string;
}

export interface TradingAccountInput {
  platformName: string;
  accountId: string;
  /** @nullable */
  accountPassword?: string | null;
  capital: number;
  profitTarget: number;
  /** @nullable */
  riskLevel?: string | null;
  /** @nullable */
  timeframe?: string | null;
}

export interface HealthStatus {
  status: string;
}

export interface BrokerInput {
  /** Broker name (e.g. Binance, Bybit, Exness, MetaTrader) */
  broker: string;
  apiKey: string;
  apiSecret: string;
  /**
     * Optional custom server URL for MT4/MT5
     * @nullable
     */
  serverUrl?: string | null;
}

export interface BrokerConnection {
  connected: boolean;
  broker: string;
  /** @nullable */
  accountId?: string | null;
  /** @nullable */
  balance?: number | null;
  /** @nullable */
  currency?: string | null;
  /** @nullable */
  connectedAt?: string | null;
}

export interface Asset {
  /** Official broker symbol (e.g. XAUUSD, BTCUSDT) */
  symbol: string;
  name: string;
  /** forex, crypto, indices, commodities */
  type: string;
  category: string;
  /** @nullable */
  spread?: number | null;
  /** @nullable */
  pip?: number | null;
}

export interface Signal {
  symbol: string;
  /** BUY or SELL */
  direction: string;
  entry: number;
  stopLoss: number;
  takeProfit1: number;
  takeProfit2: number;
  takeProfit3: number;
  timeframe: string;
  /** Confidence percentage 0-100 */
  confidence: number;
  riskPercent: number;
  /**
     * Brief analysis reason for the signal
     * @nullable
     */
  reason?: string | null;
  /** @nullable */
  generatedAt?: string | null;
}

export interface SignalRecord {
  id: number;
  symbol: string;
  direction: string;
  entry: number;
  stopLoss: number;
  takeProfit1: number;
  takeProfit2?: number;
  takeProfit3?: number;
  timeframe: string;
  confidence: number;
  /** active, won, lost, closed */
  status: string;
  riskPercent?: number;
  /** @nullable */
  profitLoss?: number | null;
  generatedAt: string;
}

export interface SignalStats {
  totalSignals: number;
  activeSignals: number;
  winRate: number;
  avgConfidence: number;
  estimatedProfit: number;
  totalWon: number;
  totalLost: number;
}

export interface EmotionStat {
  emotion: string;
  totalEntries: number;
  linkedPositions: number;
  winCount: number;
  lossCount: number;
  /** @nullable */
  winRate?: number | null;
  /** @nullable */
  avgPnl?: number | null;
}

export type PsychologyAnalyticsEntryTypeBreakdown = {[key: string]: number};

export interface PsychologyAnalytics {
  emotionStats: EmotionStat[];
  /** @nullable */
  topWinEmotion?: string | null;
  /** @nullable */
  topLossEmotion?: string | null;
  insights: string[];
  entryTypeBreakdown: PsychologyAnalyticsEntryTypeBreakdown;
  totalEntries: number;
  totalPositionsWithJournal: number;
}

export interface JournalEntry {
  id: number;
  positionId: number;
  /** open, close, note */
  entryType: string;
  /**
     * CONFIDENT, DISCIPLINED, NERVOUS, FOMO, PATIENT, GREEDY, FEARFUL, NEUTRAL
     * @nullable
     */
  emotion?: string | null;
  /**
     * Why the trader entered or is noting this
     * @nullable
     */
  reasoning?: string | null;
  /**
     * Free-form personal notes
     * @nullable
     */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface JournalEntryInput {
  positionId: number;
  entryType: string;
  /** @nullable */
  emotion?: string | null;
  /** @nullable */
  reasoning?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface PortfolioPosition {
  id: number;
  /** @nullable */
  signalId?: number | null;
  symbol: string;
  direction: string;
  entry: number;
  stopLoss: number;
  takeProfit1: number;
  takeProfit2: number;
  takeProfit3: number;
  timeframe: string;
  confidence: number;
  riskPercent: number;
  lotSize: number;
  capitalAtOpen: number;
  /** open, closed_tp1, closed_tp2, closed_tp3, closed_sl, closed_manual */
  status: string;
  /** @nullable */
  closeReason?: string | null;
  /** @nullable */
  closePrice?: number | null;
  /** @nullable */
  realizedPnl?: number | null;
  /** @nullable */
  unrealizedPnl?: number | null;
  /**
     * TP1, TP2, or TP3
     * @nullable
     */
  tpHit?: string | null;
  /** @nullable */
  isPartialClose?: boolean | null;
  openedAt: string;
  /** @nullable */
  closedAt?: string | null;
}

export interface OpenPositionInput {
  signalId: number;
  /**
     * Optional lot size override; computed from risk% if omitted
     * @nullable
     */
  lotSize?: number | null;
}

export interface ClosePositionInput {
  /** TP1, TP2, TP3, SL, MANUAL */
  closeReason: string;
  closePrice: number;
}

export interface EquityPoint {
  date: string;
  equity: number;
  pnl: number;
  positionsClosed?: number;
}

export interface PortfolioSummary {
  startingCapital: number;
  currentEquity: number;
  totalPnl: number;
  totalPnlPercent: number;
  profitTarget: number;
  profitTargetPercent: number;
  /** 0-100 percent of the way to profit target */
  targetProgress: number;
  openPositions: number;
  closedPositions: number;
  winRate: number;
  /** Average risk-reward ratio achieved on closed trades */
  avgRr: number;
  bestTrade: number;
  worstTrade: number;
  equityCurve: EquityPoint[];
}

export interface TradingProfileInput {
  capital: number;
  /** Target profit amount in dollars */
  profitTarget: number;
  /**
     * Profit target as percentage (e.g. 20 for 20%)
     * @nullable
     */
  profitTargetPercent?: number | null;
  /** LOW, MEDIUM, HIGH */
  riskLevel: string;
  /** M1, M5, M15, H1, H4 */
  timeframe: string;
}

export interface TradingProfile {
  id: number;
  capital: number;
  profitTarget: number;
  /** @nullable */
  profitTargetPercent?: number | null;
  riskLevel: string;
  timeframe: string;
  /** @nullable */
  broker?: string | null;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}