# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Merced's Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
...
components:
  schemas:
    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthUser:
      type: object
      required: [id, email, hasAccount]
      properties:
        id:
          type: integer
        email:
          type: string
        hasAccount:
          type: boolean
...